// src/types/types.ts
export interface Ssvv {
    someValue: any;
    FC: string;
    PA: string;
    SpO2: string;
    Borg_D: string;
    Borg_F: string;
    EVA_Desc: string;
}

export interface AuscultaPulmonar {
    mv: {
        abolido: boolean;
        presente: boolean;
        reduzido: boolean;
    };
    ruídos: {
        roncos: boolean;
        estridor: boolean;
        espástica: boolean;
        estertores: boolean;
        sibilos_expiratorios: boolean;
        sibilos_inspiratorios: boolean;
    };
    localizacao: {
        AHT: boolean;
        base: boolean;
        Ápice: boolean;
        direita: boolean;
        esquerda: boolean;
    };
}

export interface VitalSigns {
    id: number;
    created_at: string;
    ssvv_inicial: Ssvv;
    ssvv_final: Ssvv;
    ausculta_pulmonar: AuscultaPulmonar;
}